[clang-c-adjust] avoid converting constant array exprs to &arr[0] due to non-matching size().type()
string2array was inventing a hard-coded size() expressions for the resulting
string expression. This was signedbv and mismatched the unsignedbv array size
from the frontend. Down the line, this caused gen_typecast() to convert a
declaration of the form

  char str[] = "constant"; // size() is (size_t)9

into something like

  char str[] = &"constant"[0]; // size() is (int)9

Later on in smt-conv, an FAM special case zeroed out the RHS causing problems
with the regression/cstd/memmove family of regression tests.